Make sure powerups are persistently kept
The "powerups" variable (among other things it stores the REDFLAG and
BLUEFLAG status) that is transferred from server to client as
part of the score has to be saved as part of the score persistently.

Before it was only stored in the clientinfo, which can was overwritten
after one frame.

Since the powerups are now stored in the score we can now check it
properly for the scoreboard drawing.

Additionally i made sure that the flag can't overdraw the warmup
ready/not ready symbol.

This fixes Issue #1.