Fix vote passed/failed sound bug
When a vote passes/failes a sound is played. The client only knows
that a vote is passed/failed because of the "vote passed" text that
is send from server to client. This was parsed inadequately, such
that a player could rename himself to "vote passed" and the other
players would get the sound... We circumvent this by first checking
if a player renamed himself.