Add new cvars cg_soundOption and cg_soundOptionRail
Valid values for cg_soundOption are 0-9. A value of 0 sets all
sounds to default. A value of X (with X = 1-9) loads option X
for all sounds. A sound S.wav with option X has to have the name
S_optX.wav in the pk3.

If S_optX.wav does not exist it falls back to S.wav.

cg_soundOptionRail can overwrite the sound set by cg_soundOption
for the rail. If the option does not exist it falls back to
cg_soundOption (which falls back to the default sound).