Remove double flood limit check
Any command from the client (say or other) is going through
ClientCommand and checked for flooding with G_FloodLimited.

Some of the commands were going to G_admin_cmd_check and
then checked again? This might cause erroneous flood limiting.

Additionally, the comment above the G_FloodLimited already states
that this is checking it a second time??? Maybe this was a
copy/paste error.

This may fix Issue #5