Fix prints of long map and mapcycle lists
If the maps list got to be above 1024 chars it was not printed. The
internal buffer for the string was large enough, but OA only accepts
messages up to 1024 byte to be send from server to client.

We fix this by sending the map and mapcycle lists in 1000 byte chunks.