Scoreboard improvements
Now that the time is transferred in seconds we can show proper MM:SS.
For this we always print the time with 5 chars. To make space for it we
use the same width as for the dmg characters and for consistency we also
show the ping width the same width.

This way the scoreboard looks more consistent with the name in bold and
all of the statistics with same width font.