Make sure we never go over 1024 byte per score message
The check for the message size of the score was before the header is
added. Since it checked for exactly 1024 bytes and the header can
be ~20 bytes long, it was still possible for the resulting message
to be too big. We now check for 1000 bytes insteads, this leaves
enough room for the score-header to be added.