Add flood_protected field to client commands
The flood protection would work for all client commands. While
most client commands (vote, say, etc) are supposed to be flood
protected, some are not. For example the "score" command that
requests the score to be send to the user was also flood protected.

To handle this properly we change the commands struct and add
a flood_protected field. Each command can now be individually
set to be flood protected or not.