Update scoreboard when scoreboard key is pressed down
When the scoreboard is displayed (the tab key is pressed down) we
update it once per second. This way we can see the map time timer
run up and the deaths and frags etc are updated a bit more in
real-time. Before you had to release tab and press again to get
new data.