Fix bug where random stuff is seen as teleporter
In commit d7b5d655f we added CONTENTS_TRIGGER to the trap_Trace call in
the G_RunMissle function to be able to find teleporters and add support
for shooting through them.

This caused a bug where random stuff could be seen as a teleporter and
stop a projectile mid air. This is pretty strange, maybe a "bug"
in the map file? But obviously we can just call trap_Trace again without
CONTENTS_TRIGGER for the rest of the function to fix this