Improve team game spawn behavior
In team games it will now give the spots alternating to players from
each team. If no more spawn spots are free, the players will be randomly
placed on a spot where a player of the own team is standing.

This way there can't be any nonsensical telefrags at the beginning of the
round.

Proper handling for players spawning inside team members will be
implemented in next commits