Spawn players a bit above the original map spawn point
When we push players forward from the spawn point because a teammate
already spawned there, we check if the player would get stuck in a
solid (wall or similar). On some maps the spawn point is set a little
bit to low, so the game thinks that a pushed player ist stuck in the
floor. To circumvent this we spawn the player a tiny bit above the
original spawn.

This fixes issue #11