During spawn push, first try without pushing upwards
For ra3map1-ts we had to add a tiny bit of up-movement because players
could get stuck in the floor (spawn point a bit too low?). But with
the new up-movement it was possible to get stuck in the ceiling
in pro-dcmap7 during spawn...

So we now first try without the upwards pushing (which will work on all
maps but ra3map1-ts) and aftwards add the up-movement.

After that we still try forward/backward/top forward/top backward
(see commit before this) to make sure we dont get stuck in team
member.