Add ASXE ruleset
The ASXE ruleset will be used to try out new features with a greater
number of players. New features that are well received may become part
of the standard AS ruleset. Stuff that is unpopular will be removed again.

For the first tests we add a very light version of air control. It does
not have the usual 1.25x increase in speed and additionlly the acceleration
is decreased by 50%. No other features from CPM are used. You do have a
bit of control in the air with these settings and if you are very
good at bunny hopping you can even slowly gain a bit of speed with it.