Add zoom support for spectator mode
If you are spectating a player by default you will now see if the
player zooms.

This feature can be turned on/off with the new CVAR cg_spectatorZoom.
Use 1=on and 0=off. By default cg_spectatorZoom is set to 1. If you want
to keep the old behavior and want to control the zoom yourself in
spectator mode set cg_spectatorZoom to 0.