Use FOLLOW flag instead of spectator or isDead variables
It seems like the team (red/blue/spectator) as well as the isDead
variables in the clientinfo can be out of sync with the server data.

With many players on the server a player in red/blue team that was
alive would sometimes see the zoom animation twice. The only explanation
for this is that the code thaught the player was dead or a spectator at
this time.

Instead of the two variables we use the FOLLOW flag, which is part of
the playerstate send from the server. It should always be in sync.