Make sure item and event list is backwards compatible to q3mme
We put changes in event and item lists at the end of the respective
lists to make sure that the other elements are in the same order
as in vanilla Q3. This way the demos are more compatible to to
Q3 tools like q3mme.

Many thanks to jabba for the help.