Fix weapon properties that are used by client and server
In the last commit we added CVARs for weapon properties. Unfortunately
on a full server the implementation did not work for the properties
that are used by client and server (mostly the *Rate CVARs).

To fix this we introduce a weaponProperties message that is send to the
client on connect (or map_restart etc). So the properties that are
used only server-side applied instantly and the properites that are used
on both server and client are only applied after restart now.

The properties that are used by both client and server are
g_gauntletRate, g_machinegunRate, g_machinegunSpread, g_shotgunRate,
g_shotgunCount, g_shotgunSpread, g_plasmaRate, g_lightningRate,
g_grenadeRate, g_rocketRate, g_railRate and g_bfgRate.