Implemented 2 new shotgun pellet patterns, and made it voteable.
Vanilla OA shotgun pattern is a uniform square distribution. The intended
property is to be an effective short range weapon, useless far away. The
undesired property of its distribution is that the probability of obtaining
a lopsided pellet distribution is significant, bringing a luck aspect to
the game. Other OA mods have dealt with this in different ways, e.g. FG mod
doubled the number of pellets (and halved the damage). In AS a different
method was implemented, randomising pellet positions from a fixed pattern.
The problem is that this has rendered the shotgun far too weak, changing
weapon balance even at the short range where it should excel by design.

Four different patterns are now available through variable g_sgPattern:

oa: vanilla. Uniform square distribution
as: problematic fix from AS mod
gauss: a Gaussian distribution with the same RMS as the square distrib.
circle: Uniform circular distribution

The Gaussian distribution is enjoyable and it brings an additional skill
element, for one is required to aim properly for full power, but it alters
weapon balance as it inflicts considerable damage in the near mid range.

The circular distribution fixes the shortcomings of both the vanilla and
the as distributions, with solid damage at short range and falling off
quickly with distance. Plus, it requires slightly better aim than the
vanilla distribution.