Initial tweak of through-floors splash damage, controlled by g_thrufloors cVar.
This feature was implemented as an anti-camping measure. Whether
there is or not any need to do anything about this, it works. No more
safe camping spots that rockets can reach.

Unfortunately, through-solids splash damage breaks map geometry. Badly.
This is not about a particular map or another: every single map in the
wild is broken by introducing a feature that did not exist when authors
designed their levels. A feature that takes rid, unexpectedly, of a
fundamental attribute of solid geometry, i.e. providing cover from
projectiles.

Indeed, this is likely the single most complained about and surprising
feature of AS, and for good reason: it breaks player expectations. It is
a similar situation (only much worse) than unreachable map geometry, when
nothing but an invisible wall tells players of some badly designed limit
imposed by the mapper.

Beyond breaking maps and player expectations, as it currently stands, this
feature leads to an excess of speculative shots with a huge payback. In
other words, rocket spam is encouraged and rewarded.

Fortunately, the downsides of through-floors splash can be tamed to restore
some sanity *while* keeping it's anti-camping usefulness. The solution is
as simple as reducing the amount of damage inflicted. Before this commit,
SOLIDS OFFER NO RESISTANCE whatsoever to splash damage. They might as well
not be there. No more: g_thrufloors can adopt values from 0 (no through-floors
splash) to 1.0 (full splash) and everything in between (linear scaling factor).

Still not ideal, as damage can still be inflicted up to the full splash radius,
which again, does not feel right. Very simple fix: to be revisited.