Location ping feature, as seen in another competitive OA mod, RAT.
In team games, players can bind a key to "locPing", and use the command
to place an icon where they are aiming at, as a means of team communications.

So far only a single coloured icon is used, with neither dynamic colour change
nor HUD indicators. It can be spammed-a-lot too. Subject to testing  and revision.