Removed SERVER_INFO flag from several cvars to avoid hitting char size limit.
This seems to fix the annoying "Info string length exceeded" messages in the
console, as per an_origamian's research.

The cvars edited appear benign enough to not cause problems and do not appear
to be used outside the server VM. Limited testing though, so ample room for
screwing up I guess.