locping improvements:
1. Spatialised sound
2. Proper shader for regular players and crap one for picmip users

The spatialised sound is quieter than the local sound. To ameliorate
this the position of the locping is adjusted so that it's not too far
for the purposes of sound playing. May need tweaking.

Loading the regular shader for the icon ensures a nice and smooth
scaling. OTOH, picmip users get what they deserve.