Draw HUD scores always in the RED/BLUE order, from left to right, regardless of which team the player is in.
Swapping the scores order is confusing for no good reason. This ensures everything
stays consistent all the time. Because of AS's super HUD machinery we depend on the
position and colour settings of FlagStatus_NME, FlagStatus_OWN, Score_OWN, and
Score_NME. The default.cfg file has been edited accordingly.