BUGFIX: Movement prediction for non-vq3 rulesets.
Client wasn't aware of active ruleset (vq3, as, asxe, qw, cpm) so all the features
that depart from vq3 behaviour were not predicted. This was not too noticeable for
low pings, as the server correction was quickly available. For high pings it made
things like crouch slide unusable, and double-jumps very strange, among other things.

The fix involves storing the ruleset in player state to make it accessible. The root
of the problem was helpfully explained by origamian.